Add images to the EPUBs. Resolves #5.
In the current implementation, the images are taken from the list that
Pocket returns*. Those images are downloaded using a GET reqwest, then
passed along to the EpubBuilder. Finally, the HTML is searched for the
image URLs which are replace for the path to the new files.

*Pocket probably uses something similar to readability and that's why the
list of images returned seems to lack images.